requirements.txt
flask
opencv-python
numpy
matplotlib
scipy
scikit-learn
torch
torchvision
transformers
pillow
openai